DAQ-2666 ScanParameterDialog: fix output path calculation
- PathConstructor.createFromDefaultProperty() is no longer static,
so get PathConstructor instance from InterfaceProvider.

- Fix the calculation of subdirectory for temporary files.
The previous algorithm assumed 6 elements in the path, which is not
necessarily true (and is not true in I13 dummy mode).
This version removes the "real" subdirectory suffix (as set in
gda.data.scan.datawriter.datadir.subdir.live) and replaces it with tmp.

Note that the value displayed in the GUI is only indicative:
the tomography script determines the actual output directory.

Thus, if the script is changed, this code must be updated accordingly.

Change-Id: I3648a5ecdfceab5f1dd5b0f4f9a172601c8dfe1a